mvp 1.0 elizania front

+++ src/Controller/ProduitController.php
@@ -6,7 +6,9 @@ use App\Entity\Produit;
 use App\Entity\User;
 use App\Form\ProduitType;
 use Exception;
+use App\Controller\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\File\File;
@@ -15,14 +17,16 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
-class ProduitController extends AbstractController {
+class ProduitController extends AbstractController
+{
 
     /**
-     * @Route("/produit/create/{user_id}", name="creer-produit")
+     * @Route("/produit/create", name="creer-produit")
      */
-    public function createProduit($user_id, Request $r): Response {
-        $repository = $this->getDoctrine()->getRepository(User::class);
-        $user = $repository->find($user_id);
+    public function createProduit(Request $r): Response
+    {
+
+        $user = $this->getUser();
 
         if (empty($user)) throw new NotFoundHttpException();
 
@@ -31,7 +35,7 @@ class ProduitController extends AbstractController {
         $formProduit = $this->createForm(ProduitType::class, $produit);
         $formProduit->handleRequest($r);
 
-        if($formProduit->isSubmitted() && $formProduit->isValid()) {
+        if ($formProduit->isSubmitted() && $formProduit->isValid()) {
 
             $produit->setProducteur($user);
 
@@ -52,28 +56,26 @@ class ProduitController extends AbstractController {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($produit);
             $entityManager->flush();
-             
         }
 
         return $this->render('user/user-retrieve.html.twig', [
             'user' => $user,
-            'form' => $formProduit->createView()
-        ]); 
-        
-        // $this->redirect('/user/' . $user->getId());
+            'form' => $formProduit->createView(),
+        ]);
     }
 
 
     /**
      * @Route("/produit/{id}", name="produit")
      */
-    public function afficherUnProduit($id): Response {
+    public function afficherUnProduit($id): Response
+    {
         $repository = $this->getDoctrine()->getRepository(produit::class);
         $produit = $repository->find($id);
 
         if (empty($produit)) throw new NotFoundHttpException();
 
-        return $this->render('produit/index.html.twig', [
+        return $this->render('produit.html.twig', [
             'produit' => $produit
         ]);
     }
@@ -81,7 +83,8 @@ class ProduitController extends AbstractController {
     /**
      * @Route("/produits", name="produits")
      */
-    public function afficherTousLesArticle(): Response {
+    public function afficherTousLesArticle(): Response
+    {
         $repository = $this->getDoctrine()->getRepository(Produit::class);
         $produits = $repository->findAll();
 
@@ -95,7 +98,8 @@ class ProduitController extends AbstractController {
     /**
      * @Route("/gerer-produits", name="gerer_produits")
      */
-    public function gererProduits(): Response {
+    public function gererProduits(): Response
+    {
         $repository = $this->getDoctrine()->getRepository(Produit::class);
         $produit = $repository->findAll();
 
@@ -105,51 +109,10 @@ class ProduitController extends AbstractController {
     }
 
     /**
-     * @Route("/creer-un-produit", name="creer_produit")
-     */
-    public function creerProduit(Request $r): Response {
-
-        $produit = new Produit();
-
-        $form = $this->createForm(ProduitType::class, $produit);
-
-        $form->handleRequest($r);
-
-        if (!$form->isSubmitted() || !$form->isValid()) {
-            return $this->render('produit/creer-produit.html.twig', [
-                'form' => $form->createView()
-            ]);
-        } else {
-
-            // Je vais déplacer le fichier uploadé
-
-            // On récupère l'image
-            $image = $form->get('image')->getData();
-            // On définit le nom du fichier
-            $fileName =  uniqid() . '.' . $image->guessExtension();
-
-            try {
-                // On déplace le fichier
-                $image->move($this->getParameter('produit_image_directory'), $fileName);
-            } catch (FileException $ex) {
-                $form->addError(new FormError('Une erreur est survenue pendant l\'upload du fichier : ' . $ex->getMessage()));
-                throw new Exception('File upload error');
-            }
-
-            $produit->setPhoto($fileName);
-
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($produit);
-            $em->flush();
-
-            return $this->redirect('/produit/' . $produit->getId());
-        }
-    }
-
-    /**
      * @Route("/modifier-un-produit/{id}", name="modifier_produit")
      */
-    public function modifierProduit($id, Request $r): Response {
+    public function modifierProduit($id, Request $r): Response
+    {
 
         $repo = $this->getDoctrine()->getRepository(produit::class);
         $produit = $repo->find($id);
@@ -193,16 +156,27 @@ class ProduitController extends AbstractController {
     /**
      * @Route("/supprimer-un-produit/{id}", name="produit-delete")
      */
-    public function supprimerProduit($id): Response {
+    public function supprimerProduit($id): Response
+    {
 
         $repo = $this->getDoctrine()->getRepository(Produit::class);
         $produit = $repo->find($id);
 
         if (empty($produit)) throw new NotFoundHttpException();
 
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($produit);
-        $em->flush();
+        $user_id = $this->getUser()->getId();
+        if ($produit->getProducteur() == $user_id || $this->IsGranted('ROLE_ADMIN')) {
+
+            $filename = $produit->getPhoto();
+            // Je crée une instance de kla classe fileSystem
+            $fileSystem = new Filesystem();
+            //Je supprime l'image du dossier
+            $fileSystem->remove('%kernel.project_dir%/public/assets/photos/' . $filename);
+
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($produit);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('user-retrieve');
     }
